requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.8.3
click==8.3.0
distro==1.9.0
fastapi==0.117.1
fastapi-cli==0.0.13
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==1.109.1
outcome==1.3.0.post0
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
pysocks==1.7.1
python-dotenv==1.1.1
pyyaml==6.0.3
rich==14.1.0
rich-toolkit==0.15.1
selenium==4.35.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.19.2
typing-extensions==4.14.1
typing-inspection==0.4.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
